crm/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.5
gunicorn==20.1.0